requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
watchdog==2.1.9
